+++ bundles/fmi/language/org.eclipse.papyrus.moka.fmi.profile/src/org/eclipse/papyrus/moka/fmi/profile/custom/CustomIndependant.java
@@ -56,9 +56,9 @@ public class CustomIndependant extends IndependentImpl {
 	}
 
 	@Override
-	public int getValueReference() {
+	public long getValueReference() {
 		if (getFmiVariable() != null){
-			return (int)getFmiVariable().getValueReference();
+			return getFmiVariable().getValueReference();
 		}
 		return super.getValueReference();
 	}
